mcp_server_qdrant_astra/requirements.txt
mcp>=1.2.0
qdrant-client>=1.9.1
python-dotenv>=1.0.1
requests>=2.32.3
